feat(menu): 메뉴 트리 성능 개선 및 동기화 기능 정비
- path_id 기반으로 하위 메뉴 트리 조회 방식 전환
  - 기존 재귀적 buildMenuTree 방식 제거
  - findAllDescendantsByPathId()로 하위 메뉴 일괄 조회
  - getMenuTreeByName, getMenuTreeLiteByName 모두 path_id 방식으로 개선
  - 트리 재구성은 in-memory에서 map으로 처리하여 성능 개선

- 메뉴 트리 동기화 기능 구현
  - PUT /menu/{name}/tree/sync API 추가
  - MenuRequest 기반 트리 동기화 로직 구현 (insert/update/delete 처리)
  - syncRecursive 메서드로 재귀 트리 처리 및 path_id 갱신 처리

- MenuMapper 확장
  - toTree(List<Menu>) 메서드 추가 → flat한 Menu 리스트를 트리 DTO로 재조립
  - toFullResponse, toLiteTreeResponse 등의 커스텀 매핑 포함

- 성능 최적화를 위한 인덱스 및 제약 조건 추가
  - path_id 컬럼에 인덱스 추가 (LIKE 쿼리 최적화)
  - parent_id + position 인덱스 추가
  - type + name 조합에 유니크 키(uq_type_name) 추가

- 기존 중복 데이터 정리
  - 중복된 (type, name) 조합 총 27건 대상 name 뒤에 `_1` 접미어 추가하여 충돌 해소

+++ src/main/java/kr/co/itid/cms/entity/cms/core/menu/Menu.java
@@ -21,7 +21,7 @@ public class Menu {
     private Long parentId;  // 상속메뉴
 
     @Column(name = "position", nullable = false)
-    private Long position;  // 트리 포지션
+    private int position;  // 트리 포지션
 
     @Column(name = "level", nullable = false)
     private Long level;  // 트리 레벨
